+++ test/sdp/DefaultSDP.test.ts
@@ -1,4 +1,4 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 import * as chai from 'chai';
@@ -169,53 +169,6 @@ describe('DefaultSDP', () => {
       const sdpA = new DefaultSDP(SDPMock.VIDEO_HOST_AUDIO_ANSWER);
       const sdpB = sdpA.withBundleAudioVideo();
       expect(sdpB.sdp).to.equal(SDPMock.VIDEO_HOST_AUDIO_ANSWER_WITH_BUNDLE_AUDIO_VIDEO);
-    });
-  });
-
-  describe('withPlanBSimulcast', () => {
-    it('returns same SDP if SDP is garbage', () => {
-      const sdpLocalOffer = new DefaultSDP('garbage data');
-      const simulcastedSDP = new DefaultSDP(sdpLocalOffer.sdp).withPlanBSimulcast(2);
-      expect(simulcastedSDP.sdp).to.equal(sdpLocalOffer.sdp);
-    });
-
-    it('returns same SDP if SDP has no video section', () => {
-      const simulcastedSDP = new DefaultSDP(SDPMock.VIDEO_HOST_AUDIO_ANSWER).withPlanBSimulcast(2);
-      expect(simulcastedSDP.sdp).to.equal(SDPMock.VIDEO_HOST_AUDIO_ANSWER);
-    });
-
-    it('returns same SDP if only one layer is wanted for video', () => {
-      const sdpLocalOffer = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO);
-      const simulcastedSDP = new DefaultSDP(sdpLocalOffer.sdp).withPlanBSimulcast(1);
-      expect(simulcastedSDP.sdp).to.equal(sdpLocalOffer.sdp);
-    });
-
-    it('returns same SDP if video section has no ssrc attribute', () => {
-      const sdpLocalOffer = new DefaultSDP(
-        SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO_WITH_NO_SSRC_ATTRIBUTE_VALUE
-      );
-      const simulcastedSDP = new DefaultSDP(sdpLocalOffer.sdp).withPlanBSimulcast(2);
-      expect(simulcastedSDP.sdp).to.equal(sdpLocalOffer.sdp);
-    });
-
-    it('returns same SDP if video section has no FID group', () => {
-      const sdpLocalOffer = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO_WITHOUT_FID);
-      const simulcastedSDP = new DefaultSDP(sdpLocalOffer.sdp).withPlanBSimulcast(2);
-      expect(simulcastedSDP.sdp).to.equal(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO_WITHOUT_FID);
-    });
-
-    it('returns same SDP if original SDP only has recv video section', () => {
-      const simulcastedSDP = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_RECV_VIDEO).withPlanBSimulcast(
-        2
-      );
-      expect(simulcastedSDP.sdp).to.equal(SDPMock.LOCAL_OFFER_WITH_RECV_VIDEO);
-    });
-
-    it('returns simulcasted SDP with 2 layers if SDP has correct PlanB sendrecv video section', () => {
-      const simulcastedSDP = new DefaultSDP(
-        SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO
-      ).withPlanBSimulcast(2).sdp;
-      expect(simulcastedSDP).to.equal(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO_SIMULCAST_TWO_LAYERS);
     });
   });