+++ test/sdp/SDPMock.ts
@@ -1,4 +1,4 @@
-// Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 export default class SDPMock {
@@ -1330,5 +1330,153 @@ a=rtpmap:126 H264/90000\r
 a=rtpmap:97 H264/90000\r
 a=setup:actpass\r
 a=ssrc:4211290897 cname:{9faef5f4-f304-c442-a84c-7744cc149ec8}\r
+`;
+
+  static readonly SAFARI_AV_SENDING = `
+v=0\r
+o=- 6724755849797858675 2 IN IP4 127.0.0.0\r
+s=-\r
+t=0 0\r
+a=group:BUNDLE audio video\r
+a=msid-semantic: WMS d9a63ff3-9dcd-490c-b4a5-464f3417e812\r
+m=audio 9 UDP/TLS/RTP/SAVPF 111 103 104 9 0 8 106 105 13 110 112 113 126\r
+c=IN IP4 0.0.0.0\r
+a=rtcp:9 IN IP4 0.0.0.0\r
+a=ice-ufrag:rVEC\r
+a=ice-pwd:NoQ1g+LmzdMFxjmifwQ5aKOz\r
+a=ice-options:trickle\r
+a=fingerprint:sha-256 87:8B:3A:57:A4:8B:72:67:02:B1:B4:F9:13:72:24:EE:63:E6:69:38:70:F2:DA:68:BC:90:50:9D:27:E1:EC:65\r
+a=setup:actpass\r
+a=mid:audio\r
+a=extmap:1 urn:ietf:params:rtp-hdrext:ssrc-audio-level\r
+a=extmap:2 http://www.webrtc.org/experiments/rtp-hdrext/abs-send-time\r
+a=extmap:3 http://www.ietf.org/id/draft-holmer-rmcat-transport-wide-cc-extensions-01\r
+a=sendrecv\r
+a=rtcp-mux\r
+a=rtpmap:111 opus/48000/2\r
+a=rtcp-fb:111 transport-cc\r
+a=fmtp:111 minptime=10;useinbandfec=1\r
+a=rtpmap:103 ISAC/16000\r
+a=rtpmap:104 ISAC/32000\r
+a=rtpmap:9 G722/8000\r
+a=rtpmap:0 PCMU/8000\r
+a=rtpmap:8 PCMA/8000\r
+a=rtpmap:106 CN/32000\r
+a=rtpmap:105 CN/16000\r
+a=rtpmap:13 CN/8000\r
+a=rtpmap:110 telephone-event/48000\r
+a=rtpmap:112 telephone-event/32000\r
+a=rtpmap:113 telephone-event/16000\r
+a=rtpmap:126 telephone-event/8000\r
+a=ssrc:4085587676 cname:Zd1FwZ/qpoybYNVB\r
+a=ssrc:4085587676 msid:d9a63ff3-9dcd-490c-b4a5-464f3417e812 96b453e9-c6a9-4eed-86e0-8d18dd52b5c8\r
+a=ssrc:4085587676 mslabel:d9a63ff3-9dcd-490c-b4a5-464f3417e812\r
+a=ssrc:4085587676 label:96b453e9-c6a9-4eed-86e0-8d18dd52b5c8\r
+m=video 54073 UDP/TLS/RTP/SAVPF 96 97 98 99 100 101 102 122 127 121 125 107 108 109 124 120 123 119 114 115 116\r
+c=IN IP4 10.2.21.246\r
+a=rtcp:9 IN IP4 0.0.0.0\r
+a=candidate:249905494 1 udp 41885695 10.2.21.246 54073 typ relay raddr 0.0.0.0 rport 0 generation 0 network-id 1\r
+a=ice-ufrag:rVEC\r
+a=ice-pwd:NoQ1g+LmzdMFxjmifwQ5aKOz\r
+a=ice-options:trickle\r
+a=fingerprint:sha-256 87:8B:3A:57:A4:8B:72:67:02:B1:B4:F9:13:72:24:EE:63:E6:69:38:70:F2:DA:68:BC:90:50:9D:27:E1:EC:65\r
+a=setup:actpass\r
+a=mid:video\r
+a=extmap:14 urn:ietf:params:rtp-hdrext:toffset\r
+a=extmap:2 http://www.webrtc.org/experiments/rtp-hdrext/abs-send-time\r
+a=extmap:13 urn:3gpp:video-orientation\r
+a=extmap:3 http://www.ietf.org/id/draft-holmer-rmcat-transport-wide-cc-extensions-01\r
+a=extmap:5 http://www.webrtc.org/experiments/rtp-hdrext/playout-delay\r
+a=extmap:6 http://www.webrtc.org/experiments/rtp-hdrext/video-content-type\r
+a=extmap:7 http://www.webrtc.org/experiments/rtp-hdrext/video-timing\r
+a=extmap:8 http://tools.ietf.org/html/draft-ietf-avtext-framemarking-07\r
+a=extmap:9 http://www.webrtc.org/experiments/rtp-hdrext/color-space\r
+a=recvonly\r
+a=rtcp-mux\r
+a=rtcp-rsize\r
+a=rtpmap:96 VP8/90000\r
+a=rtcp-fb:96 goog-remb\r
+a=rtcp-fb:96 transport-cc\r
+a=rtcp-fb:96 ccm fir\r
+a=rtcp-fb:96 nack\r
+a=rtcp-fb:96 nack pli\r
+a=rtpmap:97 rtx/90000\r
+a=fmtp:97 apt=96\r
+a=rtpmap:98 VP9/90000\r
+a=rtcp-fb:98 goog-remb\r
+a=rtcp-fb:98 transport-cc\r
+a=rtcp-fb:98 ccm fir\r
+a=rtcp-fb:98 nack\r
+a=rtcp-fb:98 nack pli\r
+a=fmtp:98 profile-id=0\r
+a=rtpmap:99 rtx/90000\r
+a=fmtp:99 apt=98\r
+a=rtpmap:100 VP9/90000\r
+a=rtcp-fb:100 goog-remb\r
+a=rtcp-fb:100 transport-cc\r
+a=rtcp-fb:100 ccm fir\r
+a=rtcp-fb:100 nack\r
+a=rtcp-fb:100 nack pli\r
+a=fmtp:100 profile-id=2\r
+a=rtpmap:101 rtx/90000\r
+a=fmtp:101 apt=100\r
+a=rtpmap:102 H264/90000\r
+a=rtcp-fb:102 goog-remb\r
+a=rtcp-fb:102 transport-cc\r
+a=rtcp-fb:102 ccm fir\r
+a=rtcp-fb:102 nack\r
+a=rtcp-fb:102 nack pli\r
+a=fmtp:102 level-asymmetry-allowed=1;packetization-mode=1;profile-level-id=42001f\r
+a=rtpmap:122 rtx/90000\r
+a=fmtp:122 apt=102\r
+a=rtpmap:127 H264/90000\r
+a=rtcp-fb:127 goog-remb\r
+a=rtcp-fb:127 transport-cc\r
+a=rtcp-fb:127 ccm fir\r
+a=rtcp-fb:127 nack\r
+a=rtcp-fb:127 nack pli\r
+a=fmtp:127 level-asymmetry-allowed=1;packetization-mode=0;profile-level-id=42001f\r
+a=rtpmap:121 rtx/90000\r
+a=fmtp:121 apt=127\r
+a=rtpmap:125 H264/90000\r
+a=rtcp-fb:125 goog-remb\r
+a=rtcp-fb:125 transport-cc\r
+a=rtcp-fb:125 ccm fir\r
+a=rtcp-fb:125 nack\r
+a=rtcp-fb:125 nack pli\r
+a=fmtp:125 level-asymmetry-allowed=1;packetization-mode=1;profile-level-id=42e01f\r
+a=rtpmap:107 rtx/90000\r
+a=fmtp:107 apt=125\r
+a=rtpmap:108 H264/90000\r
+a=rtcp-fb:108 goog-remb\r
+a=rtcp-fb:108 transport-cc\r
+a=rtcp-fb:108 ccm fir\r
+a=rtcp-fb:108 nack\r
+a=rtcp-fb:108 nack pli\r
+a=fmtp:108 level-asymmetry-allowed=1;packetization-mode=0;profile-level-id=42e01f\r
+a=rtpmap:109 rtx/90000\r
+a=fmtp:109 apt=108\r
+a=rtpmap:124 H264/90000\r
+a=rtcp-fb:124 goog-remb\r
+a=rtcp-fb:124 transport-cc\r
+a=rtcp-fb:124 ccm fir\r
+a=rtcp-fb:124 nack\r
+a=rtcp-fb:124 nack pli\r
+a=fmtp:124 level-asymmetry-allowed=1;packetization-mode=1;profile-level-id=4d0032\r
+a=rtpmap:120 rtx/90000\r
+a=fmtp:120 apt=124\r
+a=rtpmap:123 H264/90000\r
+a=rtcp-fb:123 goog-remb\r
+a=rtcp-fb:123 transport-cc\r
+a=rtcp-fb:123 ccm fir\r
+a=rtcp-fb:123 nack\r
+a=rtcp-fb:123 nack pli\r
+a=fmtp:123 level-asymmetry-allowed=1;packetization-mode=1;profile-level-id=640032\r
+a=rtpmap:119 rtx/90000\r
+a=fmtp:119 apt=123\r
+a=rtpmap:114 red/90000\r
+a=rtpmap:115 rtx/90000\r
+a=fmtp:115 apt=114\r
+a=rtpmap:116 ulpfec/90000\r
 `;
 }